tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-padding-block-inline-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding-block, scroll-padding-inline with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#padding-longhands-logical">
<meta name="assert" content="scroll-padding-block, scroll-padding-inline supports the full grammar '[ auto | <length-percentage> ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-padding-block-start", "auto");
test_valid_value("scroll-padding-block-start", "10px");
test_valid_value("scroll-padding-block-start", "20%");
test_valid_value("scroll-padding-block-start", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block-start", "calc(50% + 60px)");

test_valid_value("scroll-padding-block-end", "auto");
test_valid_value("scroll-padding-block-end", "10px");
test_valid_value("scroll-padding-block-end", "20%");
test_valid_value("scroll-padding-block-end", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block-end", "calc(50% + 60px)");

test_valid_value("scroll-padding-inline-start", "auto");
test_valid_value("scroll-padding-inline-start", "10px");
test_valid_value("scroll-padding-inline-start", "20%");
test_valid_value("scroll-padding-inline-start", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline-start", "calc(50% + 60px)");

test_valid_value("scroll-padding-inline-end", "auto");
test_valid_value("scroll-padding-inline-end", "10px");
test_valid_value("scroll-padding-inline-end", "20%");
test_valid_value("scroll-padding-inline-end", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline-end", "calc(50% + 60px)");


test_valid_value("scroll-padding-block", "auto");
test_valid_value("scroll-padding-block", "10px");
test_valid_value("scroll-padding-block", "20%");
test_valid_value("scroll-padding-block", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block", "calc(50% + 60px)");
test_valid_value("scroll-padding-block", "1px 2px");
test_valid_value("scroll-padding-block", "1px auto");
test_valid_value("scroll-padding-block", "auto auto", "auto");

test_valid_value("scroll-padding-inline", "auto");
test_valid_value("scroll-padding-inline", "10px");
test_valid_value("scroll-padding-inline", "20%");
test_valid_value("scroll-padding-inline", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline", "calc(50% + 60px)");
test_valid_value("scroll-padding-inline", "1px 2px");
test_valid_value("scroll-padding-inline", "1px auto");
test_valid_value("scroll-padding-inline", "auto auto", "auto");
</script>
</body>
</html>
